+++ src/main/java/codesquad/Main.java
@@ -1,8 +1,8 @@
 package codesquad;
 
-import codesquad.http.HttpServer;
 import codesquad.http.MIME;
-import codesquad.http.handler.CreateUserHandler;
+import codesquad.server.Server;
+import codesquad.server.handlers.CreateUserHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -16,7 +16,7 @@ public class Main {
 
     public static void main(String[] args) {
         MIME.init();
-        HttpServer server = new HttpServer(PORT, THREAD_POOL_SIZE);
+        Server server = new Server(PORT, THREAD_POOL_SIZE);
 
         server.get("/create", CreateUserHandler::createUser);
         server.staticFiles("/", "/static");
