+++ src/main/java/codesquad/http/handler/Handler.java
@@ -1,5 +1,7 @@
 package codesquad.http.handler;
 
+import codesquad.http.Context;
+
 @FunctionalInterface
 public interface Handler {
     void handle(Context ctx);
