+++ src/main/java/codesquad/server/handlers/CreateUserHandler.java
@@ -1,10 +1,13 @@
-package codesquad.http.handler;
+package codesquad.server.handlers;
 
+import codesquad.http.Context;
 import codesquad.http.HttpStatus;
 import codesquad.model.User;
 import codesquad.utils.JsonConverter;
 
 public class CreateUserHandler {
+    private CreateUserHandler() {
+    }
 
     public static void createUser(Context ctx) {
         User user = new User(ctx.request().getQuery("userId"), ctx.request().getQuery("password"), ctx.request().getQuery("name"));
